Detect build-tools version for sdkmanager list
